requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
typing_extensions==4.13.2
yt-dlp==2025.3.31
